+++ app/src/main/java/com/gallery/kakaogallery/presentation/extension/ContextExtension.kt
@@ -1,0 +1,15 @@
+package com.gallery.kakaogallery.presentation.extension
+
+import android.content.Context
+import android.view.View
+import android.view.inputmethod.InputMethodManager
+import android.widget.Toast
+
+fun Context.hideKeyboard(view: View){
+    val imm = applicationContext.getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager
+    imm.hideSoftInputFromWindow(view.windowToken, 0)
+}
+
+fun Context.showToast(message: String){
+    Toast.makeText(applicationContext, message, Toast.LENGTH_SHORT).show()
+}
