+++ app/src/main/java/com/gallery/kakaogallery/presentation/extension/ViewExtension.kt
@@ -1,0 +1,8 @@
+package com.gallery.kakaogallery.presentation.extension
+
+import android.view.View
+import com.google.android.material.snackbar.Snackbar
+
+fun View.showSnackBar(message: String){
+    Snackbar.make(this, message, Snackbar.LENGTH_SHORT).show()
+}
